% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrApption.R
\name{createContrApption}
\alias{createContrApption}
\title{<Add Title>}
\usage{
createContrApption(
  counts,
  annotation,
  deResults = NULL,
  plotName = "ContrApption",
  yAxisName = "Expression",
  targetCol = "gene",
  sampleCol = "sampleID"
)
}
\description{
<Add Description>
}
