% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrApption.R
\name{ContrApption}
\alias{ContrApption}
\title{<Add Title>}
\usage{
ContrApption(
  data,
  annotation,
  groupCol,
  width = NULL,
  height = NULL,
  elementId = NULL,
  log2 = FALSE,
  plotName = NULL
)
}
\description{
<Add Description>
}
